# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Pothos API
  description: Self-hostable budget and expense tracking application API.
  version: 1.0.0
servers:
  - url: /api/v1
    description: API v1
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: pth_<64-char-hex>
...
security:
  - sessionCookie: []
  - bearerAuth: []
paths:
...
  /reports/overview:
    get:
      summary: Monthly income, expenses, and net
      tags: [Reports]
      parameters:
        - name: month
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 12
          description: Defaults to current month
        - name: year
          in: query
          schema:
            type: integer
          description: Defaults to current year
      responses:
        "200":
          description: Monthly overview. Transfers excluded.
          content:
            application/json:
              schema:
                type: object
                required: [month, year, income, expenses, net]
                properties:
                  month:
                    type: integer
                  year:
                    type: integer
                  income:
                    type: integer
                    description: Sum of income transactions (positive)
                  expenses:
                    type: integer
                    description: Absolute value of expense transactions
                  net:
                    type: integer
                    description: income - expenses (signed)
  /reports/categories:
    get:
      summary: Expense breakdown by category for a month
      tags: [Reports]
      parameters:
        - name: month
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 12
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Expenses grouped by category. Transfers excluded.
          content:
            application/json:
              schema:
                type: object
                required: [month, year, data]
                properties:
                  month:
                    type: integer
                  year:
                    type: integer
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        categoryId:
                          type: [string, "null"]
                        categoryName:
                          type: string
                          description: '"Uncategorised" if null'
                        categoryIcon:
                          type: [string, "null"]
                        categoryColor:
                          type: [string, "null"]
                        total:
                          type: integer
                          description: Absolute value (positive)
  /reports/trends:
    get:
      summary: Monthly trends over the last N months
      tags: [Reports]
      parameters:
        - name: months
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 24
            default: 12
      responses:
        "200":
          description: Monthly income/expense/net going backwards from current month. Zero-transaction months included. Transfers excluded.
          content:
            application/json:
              schema:
                type: object
                required: [months, data]
                properties:
                  months:
                    type: integer
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        month:
                          type: integer
                        year:
                          type: integer
                        income:
                          type: integer
                        expenses:
                          type: integer
                        net:
                          type: integer

  # ── Email Integration ─────────────────────────────────────────────────────
...